do not remove .gitkeep file while building static sites

+++ src/Commands/BuildStaticPagesCommand.php
@@ -92,6 +92,10 @@ class BuildStaticPagesCommand
             if ($fileInfo->isDir()) {
                 rmdir($filename);
             } else {
+                if ($fileInfo->getFilename() === '.gitkeep') {
+                    continue;
+                }
+
                 unlink($filename);
             }
         }
